introduce new human_readable_tag_name method
This is preparation work for allowing users to push tags outside
of the standard refs/tags/.* namespace. For that, the intent is to
use the AbstractUpdate.human_readable_ref_name method instead of
using the short_ref_name when printing the tag name, with 'refs/tags'
being the default_namespace.

To make this functionality more convenient to use for tags, we introduce
a new class called AbstractTagUpdate, inheriting from from AbstractUpdate
and providing a new method called human_readable_tag_name.

As a bonus, this makes the printing of tag names completely consistent
across the various types of tag updates, but also consistent with
branch updates, as shown by the changes in output in the various
tests we had to modify.

Change-Id: I8b0909001cc751d576a4c9cb4d17545acedb3462
TN: T209-003