GIT_CONFIG_OPTS: consistently use a tuple as default for tuple options
This commit is a simple cleanup commit, so as to make sure that,
for all options listed has being a "tuple" option, the default
value for the option is actually a tuple.

In practice, the code is such that it doesn't matter (it works with
None, or the empty string). But it's not the cleanest approach,
and there is no reason for it. It's just as easy as using an empty
tuple.

Doing so actually reduces coverage by one line, inside function
"to_tuple" in type_conversion.py. The uncovered code is actually
useful outside of in the handling of the default values extracted
from the GIT_CONFIG_OPTS dict. So, a test is added specifically
in order to help cover this line of code as well.

Seen while working on T512-052

Change-Id: I7e9a09c5514b29c18cdf7d92f2c5d3eabf0fe82b