replace hooks.bcc-file-ci (true/false) by hooks.filer-email (tuple)
This commit puts in place the necessary infrastructure to replace
replace the hooks.bcc-file-ci configuration in favor of a new
configuration option called hooks.filer-email.  The difference
is that the latter is a tuple/list.

For this commit, the default for the new option is such that, for
repositories that don't use the hooks.bcc-file-ci, nothing changes.
This default will be changed in a subsequent commit, and a comment
in the code explain why we're doing it this way.

This change affects two tests we have in the testsuite, which both
were using the hooks.bcc-file-ci config option:

  - "no_file_ci_bcc": We could have made this test pass by replacing
    the hooks.bcc-file-ci option by an empty hooks.filer-email.
    However, this would not be representative of the configuration
    of a real repository. So, instead, we just adjust the repository's
    configuration to match what it would typically look like, and
    accept (XFAIL) that this test is failing until we change
    the default for hooks.filer-email.

  - "to_file-ci_only": This test is in a similar situation as the test
    above, except that it has the additional distinction of now really
    becoming pointless. However, pointless or not, it's a test which
    requires little maintenance (if at all), so the decision was to
    keep that test and apply the same treatment as above.

A new test is also added, to verify that it's possible to provide
more than one email address in hooks.filer-email.

Change-Id: I1733e57b204c49cd189c5a6e1d3f4c47846b40d6
TN: T512-052