meta_config_is_immediate: do not use tabs for indentation in hooks config
This commit is to make the indentation in the hooks' config consitent,
using only spaces. As a result of this, the commit this testcase
is trying to push (on the refs/meta/config "branch") had to be
rebased on top of that change we're making here, thus causing
some minor SHA1 changes, and there some changes in output.

Change-Id: Id89ee4f8d479b7679091d987a2b1f016cc699b58
TN: T704-001