checker_config_on_its_own: use consistent indentation in hooks config
This testcase uses a repository which has some branches which are based
on the bare repository's refs/meta/config "branch". Thus, after having
updating the hooks configuration in that reference, we had to rebase
those branches, which means changes in metadata (in particular SHA1s),
affecting the output obtained when using those branches during
the course of the testcase.

Change-Id: If581d8739813b6446e2bacd1dbec83ba1731da41
TN: T704-001