CommitInfo: Keep the author's name and email address separate
Currently, the CommitInfo object records information about the author
of a commit as a single element, which is a string that contains
both the author's name and his email address, in RFC 822 format.
As of now, this is all the git-hooks need; however, in the context
of adding a hook that projects could use implement project-specific
validation rules, such hook might prefer to have name and email
addresses be passed as separate entities. This change makes this
possible without requiring an additional call to Git.

Note that this commit introduces a new property in class CommitInfo
with the same name ("author") as the attribute which has now been
split into two new attributes. This change highlights the fact that
the name we previously chose is very compact ("author"), but not
overly expressive. This commit keeps things as they are now to keep
the focus on the splitting. A followup patch will focus on the renaming.

Change-Id: I2b3dd4863f37c4364d4c477bcfc3952cedd2944a
TN: T209-005