rename CommitInfo's author property to full_author_email (clearer)
Initially, this property was born an attribute. It was then converted
to a property when the attribute's value got split into two attributes.
The original name of the attribute was used for this property so as
to facilitate the transition, as well as review of that change.

Now that the transition is done, this commit renames the property
into something which is more meaningful and obvious to the reader,
and then updates all the users.

Change-Id: I29f759dfae2a71a1c9b1a65fd93fe7604b6d952a
TN: T209-005