AbstractUpdate.commit_data_for_hook: New method to factorize code
This commit factorizes some code which is currently defined locally
when handling the hooks.commit-extra-checker config option. The intent
is to build on it when we add support for a new option allowing users
to customize the contents of the commit emails being sent (the info
we will want to pass to the new hook to customer emails will be
a superset of the information passed here).

Change-Id: I3cc9659b80b0b03752d227966776c804d3c5203a
TN: T209-006