Update incorrect comment in AbstractUpdate.__check_max_commit_emails
This commit updates a comment which hasn't been true anymore for
quite a while. It is written in such a way that it is neutral about
how the policy used to determine which commits trigger an email
and which don't, so as to avoid being wrong again the next time
we introduce new policies.

Change-Id: Icbbf7d5451e7ce2bf6b3fd0e4e10cdca902b38b4
TN: T815-009