not_ISO_8859-15: merge all testing into a single method
The issue with having two methods is that the order in which those
methods are executed is not visual and can create situations
where the test is mislead because the order chosen by unittest to
execute the testing methods is different from the order they appear
in the test script. For this test, as of now, this does not matter,
because nothing gets pushed to the bare repository. However, in
an upcoming change where we add a new feature allowing users to
push commits with revision logs containing characters outside
the range allowed by default, the order will matter. Prevent
this issue by merging the two methods into a single one. The downside
is that a failure in the first test will prevent the second one
from running. But since we never allow changes that fail a test
to get pushed to production, this shouldn't be an issue in practice.

Change-Id: I4f8f17b8ae0a73055c1207de62e9615500dd9b57
TN: T829-004