not_ISO_8859-15: Add bad-after-subject branch to bare repository
This commit just changes this testcase's bare repository to add
a new branch, called bad-after-subject, pointing to the same commit
as the master branch. This new branch name was chosen to match
the name of one of the two branches used by this testcase.
The testcase is then changed to try to push its bad-after-subject
branch to bad-after-subject, rather than re-using master as
the destination again.

This change doesn't make a difference at the moment, because neither
of the two push operations performed by the test are expected to
succeed. But they are preparation work for a followup change which
provides the ability to skip the verifications causing these pushes
to fail. In that followup commit, we'll have the testcase change
the test repository's configuration, and then re-attempt the pushes,
expecting them to succeed this time. When we do that, we'll want
to push each of the two branches to different branches on the remote.
That's why this commit introduces the second branch on the remote.

Change-Id: I86e0c18ae094818821202e545e646cbcad595c25
TN: T829-004