AbstractUpdate.pre_commit_checks: Remove unnecessary early return
This method computes a list of commits that need to be checked,
and then makes the assumption that all the checks should be applied
to the same list. This is mostly correct, except in the case of
merge commits: The list of commits to check excludes commits
which have a revision log suggesting that they are revert commits.
In the interest of keeping a given branch free of merge commits,
we still want to reject the revert commits that are merges. So
the list of commits to check for merge-commit violations is
different from the list used by most checks.

In practice, this early return doesn't contribute anything tangible
in terms of performance, and doesn't allow each check to decide
what list of commits it should operate on.

So this commit removes it. There should be no user-visible effect
coming from this change.

Seen while working on TA02-034.

Change-Id: I3a7e826ed4dba36f75da2fc088360356e5a8fa70