update.py crash when not commits to style-check and combined-style-checking
This commit fixes a crash which occurs during the validation phase
(when the "update" hook is called). This only happens when the repository
is configured with the combined-style-checking config option set.
So far, we found a couple of scenarios where this could happen:

  - The user pushes a single commit, which happens to be a "revert"
    commit (i.e. a commit which is created via the "git revert"
    command); those commits are purposefully excluded from the list
    of commits we have to check;

  - The user pushes a tag, where the tag points to a commit which
    already exists in the repository (something very much typical
    of people pushing tags).

In both cases, we land in the AbstractUpdate.__do_style_checks method,
doing the following:

        | added = self.commits_to_check
        | if git_config('hooks.combined-style-checking'):
        |     # This project prefers to perform the style check on
        |     # the cumulated diff, rather than commit-per-commit.
        |     # Behave as if the update only added one commit (new_rev),
        |     # with a single parent being old_rev.  If old_rev is nul
        |     # (branch creation), then use the first parent of the oldest
        |     # added commit.
        |     debug('(combined style checking)')
 !!! -> |     if not added[-1].pre_existing_p:
        |         [...]

The problem in the scenarios mentioned above is that, because there are
no commits to check, added is the empty list, and so trying to get
the last element of that list does not work.

This commit fixes the problem by checking the list of added commits
early, and returning right away if there are no commits to check.

As a result of this early return, a couple of tests saw their output
change, because a debug log is no longer printed. I verified that
these were for two tests where the push did not send any commits
that are new to the repository. So the change of behavior is expected.

Change-Id: I28e5ffe61ec19b8acf46b6da5c0bfe5ed0139d2c
TN: TA21-016