TB15-003__non_ascii_char_in_diff: New testcase.
This testcase was an attempt at trying to reproduce the problem
reported in https://github.com/AdaCore/git-hooks/issues/13.

As it happens, this testcase is not sufficient to reproduce
the problem above; but it's sufficiently interesting on its own
that it's worth adding to our testsuite nonetheless, while we continue
trying to find another reproducer for the GitHub issue above.

Change-Id: I1c63d64797ebb9ccd79eca52e45e179e741c9132
TN: TB15-003