hooks/daemon.py: Allow parsing with Python-3 version of flake8
This commit changes a couple of exception handlers to avoid using
a syntax which is no longer valid in Python 2.x, using a syntax
which is valid in both Python 2.x and Python 3.x insteadl.

This is good preparation for the transition to Python 3.x, and
this also makes sure the current version of this module now passes
AdaCore's internal style checker.

Change-Id: I80bd4ab61eb7bc3caa23a900cbdce66748f670b6