hooks/utils.py: Allow parsing with Python-3 version of flake8
This commit replaces the litteral number "0664", which is no longer
valid syntax in Python 3.x, with "0o664", which is valid with both
Python 2.x and Python 3.x.

This is good preparation for the transition to Python 3.x, and
this also makes sure the current version of this module now passes
AdaCore's internal style checker.

Change-Id: I43ea77e894d8abbf8b36dfd40da2b8a4856489fd