hooks/git.py: Replace reference to "basestring" by "str"
This type will disappear when using a Python 3.x interpreter.

In the meantime, while this code is still run using a Python 2.7
interpreter, it's not a loss in practice to use "str" instead
because we do not use this code in a context where we pass
a unicode string as the _outfile parameter.

Change-Id: Ib56c1521324ee8a8b8b89a7d3b97f43328c313bb