fix UnicodeEncodeError crash while calling file-commit-cmd hook
This commit fixes an unexpected UnicodeEncodeError error which
happens when a user pushes a change to a repository which uses
both a file-commit-cmd hook as well as a commit-email-formatter
hook which customizes commit-email bodies.

While this doesn't happen when the commit-email-formatter only
changes the email subject or the "diff", tests are also added
to cover these cases, for completeness reason.

Change-Id: I360fcaf25db6cd4a930302e2d6f34e8609e2b364
TN: U110-001