default of tuple config option ignored when used in project's config
Projects with a git-hooks configuration using options which are declared
as a "tuple/list" cause the default of such options to be ignored.
This does not really make sense, because it forces users who want
to add to the default configuration to first repeat that default
configuration.

This commit changes the behavior to make sure that the git-hooks
don't drop the default when the user uses such a config option.

This is unfortunately a bit of a compatibility break with previous
revisions. In pratice, all but one config option ("hooks.ignore-refs")
have an empty default, so it's only a potential issue for that one
option. This "hooks.ignore-refs" option was created specifically
for Gerrit support and it is believed that the chances of a project
being negatively affected by this change are very small -- and if
this happens, we will provide a mechanism for allow the default
to be ignored. In the meantime, this commit updates the documentation
to make things clearer in that respect.

Change-Id: If5ce8ab7533da21bc3190f8b8fc154b17aa85a36
TN: U227-003