testsuite: change the testcase filename from "test.py" to "run_test.py"
This is a small step towards migrating this testsuite towards
a pytest-based framework: The filename currently used for testcases
is "test.py", and that filename does not match the nomenclature
that pytest expects for testcase files. So this commit renames
all our testcases to something that pytest can identify as a testcase
file to load and execute, and then adjusts the current testsuite
framework accordingly.

Change-Id: I77fba65e8ea20612259c6ad10459fe96ef77a2fc