always send all emails using a UTF-8 charset
This commit simplifies the choice of the charset being used to send
our emails to just using UTF-8. This makes all emails consistently using
that charset, and in particular follows something we were already doing
when the message body was found to be in unicode format (this happens
when the message body comes from calling one of the project's hooks).

The expectation is that this preliminary change will facilitate
the transition to Python 3, where strings are unicode.

Change-Id: I0e44baf460dd99a2505d94671ac6042304addfd2
TN: TB22-002