Use iso-8859-15 instead of iso-8859-1 when guessing the encoding
When looking at the 8 code points that are different between
iso-8859-1 and iso-8859-15, it seems like the iso-8859-15 has
some characters might be more widely used than the ones in
iso-8859-1 (some standalone accents, some fractions, a generic
"currency sign"). This commit therefore changes the hooks to
use iso-8859-15 instead of iso-8859-1.

Note that this doesn't change the fact that UTF-8 is still
the first-choice encoding we try.

Change-Id: I36092552dc647935269b1f0f6b401d198e1a7bd6
TN: U528-040