testsuites/bin/{pack,unpack}-test-repos: Remove use of gnatpython.ex.Run
This commit removes another use of gnatpython in the testsuite, this time
eliminating the use of gnatpython.ex.Run, using the equivalent Popen class
instead. With Python 3.x, we would be able to write this a bit more simply
thanks to the "subprocess.run" function, but we are are to ready to drop
support for Python 2 yet, so we implemented this in a way that works with
both Python 2 and Python 3.

Change-Id: I9a1e9c0ece11de4ce5001d052cffc20ac97bd566
TN: U530-006