simplify "diff" section computation in commit emails
This commit is inspired by the fact that I couldn't understand
why I was skipping the first character from the output of a Git
command whe computing a commit's "diff", like so):

    diff = git.show(commit.rev, [...], pretty="format:|")[1:]

(emphasis on the "[1:]" at the end).

To understand, I remove the subscripting and reran the testsuite
without it to see what failures I would get. This gave me
the answer, which is we were intentionally starting the "format:"
string with a "|", and so we needed to strip that extra character.

That's when I found a comment I wrote; I didn't see it at first
because it was placed further up:

     # For the diff, there is one subtlelty:
     # Git commands calls strip on the output, which is usually
     # a good thing, but not in the case of the diff output.
     # Prevent this from happening by putting an artificial
     # character at the start of the format string, and then
     # by stripping it from the output.

This may have made sense back when I wrote that comment, but
we no longer strip the start of the output anymore (see commit
af06d5ea5479dc67bab342b69a4431865f29e4af). So I decided to
simplify the code by removing the extraneous character in
the "format:" string.

As it happens, this revealed that git behaves slightly differently
when given an empty "format:" string. Before:

    | $ git show -p -M --stat --pretty="format:|" HEAD
    | |---
    |  hooks/git.py | 4 +---
    |  1 file changed, 1 insertion(+), 3 deletions(-)
    |
    | diff --git a/hooks/git.py b/hooks/git.py
    | index fe2b36b..0669111 100644
    | [snip]

After (removing the "|" in the "format:" string):

    | $ git show -p -M --stat --pretty="format:" HEAD
    |  hooks/git.py | 4 +---
    |  1 file changed, 1 insertion(+), 3 deletions(-)
    |
    | diff --git a/hooks/git.py b/hooks/git.py
    | [snip]

What we can see is that the "---" separate line is no longer shown
in the second command.

Rather than forcing Git to print it, or rather than staying with
the existing code, this commit simply hardcodes the separator line.
One minor bonus of doing it this way is that, if Git decides to
change that separator, this won't affect us, and thus we won't
have to change hundreds of tests accordingly.

And by doing so, this revealed that there was actually an inconsistency
in the formatting produced by Git: In some cases (e.g. merge commits),
it became apparent that Git was omitting this "---" separator line,
even when the "format:" string was empty. The corresponding testcases
where the inconsistency showed up were adjusted to match the new
behavior, which is consisdered (slightly) better, because more
consistent.

Found while working on U530-006 (transition to Python 3.x).

Change-Id: Ifc473fa471ba618e11c3c4bcc6d83cc6f82fc6bf