check_fast_forward: Minor test rewrite to avoid string comparison
This commit modifies a conditional expression to avoid comparing
the output from a git command with the empty string. With Python 3.x,
this wouldn't work unless we decoded the output first or compared
against the empty *byte* string. But since all we're interested in
is knowing whether the output is empty or not, we can use a simple
"truth" test instead, thus avoiding having to worry about the type
of the output entirely.

TN: U530-006
Change-Id: I0067cb6bfb37da4e1782a045beac19d970d7718f