style_check_files: Allow "filename_list" param to be any kind of iterable
The "filename_list" of the style_check_files function is expected
to be a list. However, as it happens, we have a call to this function
where the filename_list parameter was computed using the "filter"
function.

    files_to_check = filter(needs_style_check_p, files_to_check)
    style_check_files(files_to_check, new_rev, project_name)

This is no problem when trying to iterate on the list alone, but
at one point inside the "style_check_files" funtion, we try to
add to it like so:

    for filename in filename_list + aux_files:

With Python 3.x, this causes a crash because the "filter" function
now returns a generator rather than a list. Seems it seems desirable
to allow "style_check_files" to accept any iterable as "filename_list",
this commit modifies the code to allow it.

TN: U530-006
Change-Id: I8e5caaf09d177096788c522d7bff8168eb158ec7