style_check_commit: Convert to tuple return value from call to "filter"
This is a preparation patch for the transition to Python 3.x.

Currently, we have the following code:

    | files_to_check = filter(needs_style_check_p, files_to_check)
    | if not files_to_check:
    |     debug("style_check_commit: no files to style-check")
    |     return

With Python 2.x, the "filter" function constructs a list, so
we can indeed check for emptiness using the logical operator "not".

With Python 3.x, however, filter returns an iterator, and using
the logical operator "not" no longer has the same semantics at all.

The easiest way to handle this is to simply convert the result to
a tuple.

TN: U530-006
Change-Id: I334f2657728405226d7da577e893b7c64a3220fb