testsuite/coverage.sh: Use coverage.rc during report generation
The coverage.sh script creates a coverage.rc file, but this file wasn't
actually used during coverage generation. This change makes sure that
it is.

For now, this shouldn't make any difference, but this will help us
later for providing additional options during report generation
(in particular, it will allow us to add coverage exclusion patterns,
which will become handy to exclude code which is only meant to be
executed with a given version of Python).

TN: U530-006
Change-Id: Ifbba6dfb9a9347f7d5680b6cc464378282653813